+++ common/src/main/java/dev/felnull/imp/IMPHoliday.java
@@ -11,7 +11,16 @@ public class IMPHoliday {
         return calendar.get(Calendar.MONTH) == Calendar.DECEMBER && calendar.get(Calendar.DATE) >= 24 && calendar.get(Calendar.DATE) <= 26;
     });
 
+    private static final Supplier<Boolean> APRIL_FOOL = Suppliers.memoize(() -> {
+        Calendar calendar = Calendar.getInstance();
+        return calendar.get(Calendar.MONTH) == Calendar.APRIL && calendar.get(Calendar.DATE) == 1;
+    });
+
     public static boolean isXmas() {
         return XMAS.get();
+    }
+
+    public static boolean isAprilFool() {
+        return APRIL_FOOL.get();
     }
 }